+++ CHANGELOG.md
@@ -10,6 +10,7 @@ https://github.com/capistrano/capistrano/compare/v3.4.0...HEAD
 
 * Minor changes
   * Fix filtering behaviour when using literal hostnames in on() block (@townsen)
+  * Added options to set username and password when using Subversion as SCM (@dsthode)
 
 ## `3.4.0`